Extracting getPages and paginate from core/list into separate files
