+++ lib/list.js
@@ -888,122 +888,8 @@ List.prototype.getUniqueValue = function(path, generator, limit, callback) {
 	check();
 };
 
-/**
- * Generate page array for pagination
- *
- * @param {Number} the maximum number pages to display in the pagination
- * @param {Object} page options
- */
-List.prototype.getPages = function(options, maxPages) {
-	var surround = Math.floor(maxPages / 2);
-	var firstPage = maxPages ? Math.max(1, options.currentPage - surround) : 1;
-	var padRight = Math.max(((options.currentPage - surround) - 1) * -1, 0);
-	var lastPage = maxPages ? Math.min(options.totalPages, options.currentPage + surround + padRight) : options.totalPages;
-	var padLeft = Math.max(((options.currentPage + surround) - lastPage), 0);
-	options.pages = [];
-	firstPage = Math.max(Math.min(firstPage, firstPage - padLeft), 1);
-	for (var i = firstPage; i <= lastPage; i++) {
-		options.pages.push(i);
-	}
-	if (firstPage !== 1) {
-		options.pages.shift();
-		options.pages.unshift('...');
-	}
-	if (lastPage !== Number(options.totalPages)) {
-		options.pages.pop();
-		options.pages.push('...');
-	}
-};
-
-/**
- * Gets a special Query object that will paginate documents in the list
- *
- * Example:
- *     list.paginate({
- *         page: 1,
- *         perPage: 100,
- *         maxPages: 10
- *     }).exec(function(err, results) {
- *         // do something
- *     });
- *
- * @param {Object} options
- * @param {Function} callback (optional)
- */
-List.prototype.paginate = function(options, callback) {
-	var list = this;
-	var model = this.model;
-
-	options = options || {};
-
-	var query = model.find(options.filters);
-
-	query._original_exec = query.exec;
-	query._original_sort = query.sort;
-	query._original_select = query.select;
-
-	var currentPage = Number(options.page) || 1;
-	var resultsPerPage = Number(options.perPage) || 50;
-	var maxPages = Number(options.maxPages) || 10;
-	var skip = (currentPage - 1) * resultsPerPage;
-
-	list.pagination = { maxPages: maxPages };
-
-	// as of mongoose 3.7.x, we need to defer sorting and field selection
-	// until after the count has been executed
-
-	query.select = function() {
-		options.select = arguments[0];
-		return query;
-	};
-
-	query.sort = function() {
-		options.sort = arguments[0];
-		return query;
-	};
-
-	query.exec = function(callback) {
-		query.count(function(err, count) {
-			if (err) return callback(err);
-
-			query.find().limit(resultsPerPage).skip(skip);
-
-			// apply the select and sort options before calling exec
-			if (options.select) {
-				query._original_select(options.select);
-			}
-
-			if (options.sort) {
-				query._original_sort(options.sort);
-			}
-
-			query._original_exec(function(err, results) {
-				if (err) return callback(err);
-				var totalPages = Math.ceil(count / resultsPerPage);
-				var rtn = {
-					total: count,
-					results: results,
-					currentPage: currentPage,
-					totalPages: totalPages,
-					pages: [],
-					previous: (currentPage > 1) ? (currentPage - 1) : false,
-					next: (currentPage < totalPages) ? (currentPage + 1) : false,
-					first: skip + 1,
-					last: skip + results.length
-				};
-				list.getPages(rtn, maxPages);
-				callback(err, rtn);
-			});
-		});
-
-	};
-
-	if (callback) {
-		return query(callback);
-	} else {
-		return query;
-	}
-};
+List.prototype.getPages = require('./list/getPages');
+List.prototype.paginate = require('./list/paginate');
 
 /*!
  * Export class
